Rename the `path` module to `fileresolver`
Originally called `imagefinder`, it was renamed to `imagepath` as
`imagefinder` started referring to the module responsible for
using computer vision to locate images on a region. Later the
"image" prefix was dropped as the modules had become more
generalized, supporting other targets beyond simple images.

As a result of these renames, the name of the module stopped
conveying the correct meaning as the term "path" is too broad.
Unlike `location` and `target` it doesn't represent a path, and
although it performs an action, it doesn't end with `er` as
`imagelogger` and `finder`.

This commit renames it one more time — now to `fileresolver` — in
accordance with other similar modules. The legacy `path` module
was kept for compatibility reasons, but is now deprecated and will
be removed in later versions.
